refactor: update on session view
