+++ src/frontend/src/modals/IOModal/components/SessionView/index.tsx
@@ -1,5 +1,5 @@
 import Loading from "@/components/ui/loading";
-import { useGetMessagesQuery } from "@/controllers/API/queries/messages";
+import { useGetMessagesQuery, useUpdateMessage } from "@/controllers/API/queries/messages";
 import { useIsFetching } from "@tanstack/react-query";
 import {
   CellEditRequestEvent,
@@ -10,7 +10,6 @@ import cloneDeep from "lodash/cloneDeep";
 import { useMemo, useState } from "react";
 import TableComponent from "../../../../components/tableComponent";
 import useRemoveMessages from "../../../../pages/SettingsPage/pages/messagesPage/hooks/use-remove-messages";
-import useUpdateMessage from "../../../../pages/SettingsPage/pages/messagesPage/hooks/use-updateMessage";
 import useAlertStore from "../../../../stores/alertStore";
 import { useMessagesStore } from "../../../../stores/messagesStore";
 import { messagesSorter } from "../../../../utils/utils";
@@ -26,6 +25,7 @@ export default function SessionView({
   const messages = useMessagesStore((state) => state.messages);
   const setErrorData = useAlertStore((state) => state.setErrorData);
   const setSuccessData = useAlertStore((state) => state.setSuccessData);
+  const updateMessage = useMessagesStore((state) => state.updateMessage);
 
   const isFetching = useIsFetching({
     queryKey: ["useGetMessagesQuery"],
@@ -40,7 +40,7 @@ export default function SessionView({
     selectedRows,
   );
 
-  const { handleUpdate } = useUpdateMessage(setSuccessData, setErrorData);
+  const {  mutate:updateMessageMutation} = useUpdateMessage();
 
   function handleUpdateMessage(event: NewValueParams<any, string>) {
     const newValue = event.newValue;
@@ -50,10 +50,19 @@ export default function SessionView({
       ...row,
       [field]: newValue,
     };
-    handleUpdate(data).catch((error) => {
+    updateMessageMutation(data,{onSuccess:()=>{
+      updateMessage(data);
+      // Set success message
+      setSuccessData({
+        title: "Messages updated successfully.",
+      });
+    },onError:()=>{
+      setErrorData({
+        title: "Error updating messages.",
+      });
       event.data[field] = event.oldValue;
       event.api.refreshCells();
-    });
+    }});
   }
 
   const filteredMessages = useMemo(() => {
